Chatbot/requirements.txt
fastapi
uvicorn
langchain
langchain-community
chromadb
pypdf
openai
tiktoken
python-dotenv
